requirements.txt
discord.py
pillow
opencv-python-headless
pytesseract
python-dotenv